Clean up terminal code and move to UI module
Changes:
- Move Swing based terminal classes into the UI module
- Combine InteractiveTerm into MarsTerminal
- UIConfig uses the same structure for main window & terminal
- MarsTerminal handles positioning itself
- Increase width of update panel in splash screen
- Move textIO module from console to UI module
Ref: #356
