+++ mars-sim-console/pom.xml
@@ -20,11 +20,6 @@
 		<version>${project.version}</version>
 	</dependency>
 	<dependency>
-		<groupId>org.beryx</groupId>
-		<artifactId>text-io</artifactId>
-		<version>3.4.1</version>
-	</dependency>
-	<dependency>
 	    <groupId>org.apache.sshd</groupId>
 	    <artifactId>sshd-core</artifactId>
 	    <version>2.16.0</version>
